[TOC] - refactor. moved detection algo to be able to easily run more than one
