+++ server/src/processing/TableOfContentsDetectionModule/detection-methods.ts
@@ -1,0 +1,29 @@
+import { BoundingBox, Paragraph, Word } from './../../types/DocumentRepresentation';
+
+export const threshold = 0.4;
+
+/*
+  this detection method searches for text finishing in numbers in the right 10% width area of the BBox
+*/
+export function endsWithNumber(p: Paragraph): boolean {
+  const w = p.width * 0.1;
+  const intersectionBox = new BoundingBox(p.right - w, p.top, w, p.height);
+  const wordsInsideIntersection =
+    p.getWords()
+      .filter(word => BoundingBox.getOverlap(word.box, intersectionBox).box1OverlapProportion > 0)
+      .filter(word => !isSeparator(word));
+
+  return wordsInsideIntersection.filter(isNumber).length > Math.floor(wordsInsideIntersection.length * 0.5);
+}
+
+function isNumber(word: Word): boolean {
+  const decimalNumbers = new RegExp(/[0-9]+$/);
+  const romanNumbers = new RegExp(/^[ivxlcdmIVXLCDM]+$/);
+  const w = word.toString();
+  return decimalNumbers.test(w) || romanNumbers.test(w);
+}
+
+function isSeparator(word: Word): boolean {
+  const separators = new RegExp(/^[-. ]+$/);
+  return separators.test(word.toString().trim());
+}
